feature - EsaFilters component creation
